rename: 폴더 위치 변경 client 관련 파일들 client 폴더에서 관리
