+++ client/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="index.css" />
+    <link rel="stylesheet" href="./dist/main.css" />
     <title>TO DO LIST</title>
   </head>
   <body>
@@ -27,6 +27,6 @@
     <div class="drag"></div>
     <div class="fabBtn"><p class="fab__addBtn">+</p></div>
 
-    <script type="module" src="app.js"></script>
+    <script type="module" src="./dist/bundle.js"></script>
   </body>
 </html>
