+++ client/package.json
@@ -3,7 +3,6 @@
   "version": "1.0.0",
   "description": "TodoList",
   "main": "app.js",
-  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack --mode production"
