+++ client/style/_alarm.scss
