+++ client/style/_modal.scss
